+++ pymake/core/gram.py
@@ -23,10 +23,6 @@ class ExpArgumentParser(argparse.ArgumentParser):
         self.print_usage()
         print('error', message)
         #self.print_help()
-        try:
-            os.remove('.pmk-db.db')
-        except FileNotFoundError:
-            pass
         sys.exit(2)
 
 class VerboseAction(argparse.Action):